+++ ChangeLog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `e3nn.IrrepsArray.filtered(keep: List[Irrep])`
 - `e3nn.reduced_tensor_product_basis(formula_or_irreps_list: Union[str, List[e3nn.Irreps]], ...)`
 - `IrrepsArray.at[i].set(v)` and `IrrepsArray.at[i].add(v)`
+- add `Irreps.is_scalar`
 
 ## [0.9.0] - 2022-09-04
 ### Added